Create 'handle_operations' mod to move functional code out of 'mod.rs'
