+++ src/operations/op_not.rs
@@ -12,9 +12,9 @@ use crate::{registers::Registers, utils::update_flags, vm_error::VmError};
 /// - `instr`: A 16-bit instruction.
 ///
 /// # Returns
-/// 
+///
 /// Returns `Ok(())` if the operation was successful, otherwise returns a `VmError`.
-/// 
+///
 pub fn op_not(registers: &mut Registers, instr: u16) -> Result<(), VmError> {
     let r0 = (instr >> 9) & 0x7;
     let r1 = (instr >> 6) & 0x7;
