+++ src/operations/op_st.rs
@@ -11,9 +11,9 @@ use crate::{memory::Memory, registers::Registers, utils::sign_extend, vm_error::
 /// - `instr`: A 16-bit instruction.
 ///
 /// # Returns
-/// 
+///
 /// Returns `Ok(())` if the operation was successful, otherwise returns a `VmError`.
-/// 
+///
 pub fn op_st(registers: &mut Registers, instr: u16, memory: &mut Memory) -> Result<(), VmError> {
     let r0 = (instr >> 9) & 0x7;
     let pc_offset = sign_extend(instr & 0x1FF, 9);
