+++ src/utils.rs
@@ -10,9 +10,9 @@ use crate::vm_error::VmError;
 /// Reads an image file into memory. The image file is expected to start with
 /// a 16-bit address indicating where in memory the data should be loaded, followed by
 /// 16-bit instructions to be stored sequentially in memory.
-/// 
+///
 /// # Returns
-/// 
+///
 /// An `Ok` result if the image file reading was successful, otherwise a `VmError`.
 pub fn read_image_file(path: &str, memory: &mut Memory) -> Result<(), VmError> {
     let file = File::open(path).map_err(|e| VmError::FailedToOpenFile(e.to_string()))?;
@@ -37,11 +37,11 @@ pub fn read_image_file(path: &str, memory: &mut Memory) -> Result<(), VmError> {
 }
 
 /// Flushes the stdout buffer
-/// 
+///
 /// # Returns
-/// 
+///
 /// An `Ok` result if the operation was successful, otherwise a `VmError`.
-/// 
+///
 pub fn flush_stdout() -> Result<(), VmError> {
     std::io::stdout()
         .flush()
@@ -49,11 +49,11 @@ pub fn flush_stdout() -> Result<(), VmError> {
 }
 
 /// Updates the condition flags in the `Registers` struct.
-/// 
+///
 /// # Returns
-/// 
+///
 /// An `Ok` result if the operation was successful, otherwise a `VmError`.
-/// 
+///
 pub fn update_flags(registers: &mut Registers, r: u16) -> Result<(), VmError> {
     if registers.get(r)? == 0 {
         registers.cond = FL_ZRO;
